Address PR review feedback
- Move ConstructorInfo to separate file
- Fix EnumFieldParser parameter line formatting
- Add newline at end of TypeParser
- Remove column alignment in TypeParser javadoc
- Extract setter-based logic to addSetterBasedParseMethods method
- Extract field processing to generateFieldParsingCode method
- Use imports for List and ArrayList in Main.java
